CMD "main" package is built from repo root Can now do simple go build && ./go-transcode
