+++ build
@@ -1,3 +1,4 @@
 #!/bin/sh
 
-go build -o bin/transcode cmd/transcode/main.go
+#go build -o bin/transcode cmd/transcode/main.go
+go build
